[Ref] Splitting EC2 backend into classes, pt. 6
