+++ etc/backends/ec2/model/compute/contextualization.json
