+++ lib/backends/ec2/compute.rb
@@ -1,8 +1,12 @@
 module Backends
   module Ec2
-    module Compute
+    class Compute < Backends::Ec2::Base
       COMPUTE_NINTF_REGEXP = /compute_(?<compute_id>i-[[:alnum:]]+)_nic_(?<compute_nic_id>eni-[[:alnum:]]+)/
       COMPUTE_SLINK_REGEXP = /compute_(?<compute_id>i-[[:alnum:]]+)_disk_(?<compute_disk_id>vol-[[:alnum:]]+)/
+
+      DALLI_OS_TPL_KEY = 'ec2_os_tpls'
+      IMAGE_FILTERING_POLICIES_OWNED = ['only_owned', 'owned_and_listed'].freeze
+      IMAGE_FILTERING_POLICIES_LISTED = ['only_listed', 'owned_and_listed'].freeze
 
       # Gets all compute instance IDs, no details, no duplicates. Returned
       # identifiers must correspond to those found in the occi.core.id
@@ -418,6 +422,139 @@ module Backends
         end
 
         true
+      end
+
+      # Returns a collection of custom mixins introduced (and specific for)
+      # the enabled backend. Only mixins and actions are allowed.
+      #
+      # @return [Occi::Collection] collection of extensions (custom mixins and/or actions)
+      def compute_get_extensions
+        read_extensions 'compute', @options.model_extensions_dir
+      end
+
+      # Gets backend-specific `os_tpl` mixins which should be merged
+      # into Occi::Model of the server.
+      #
+      # @example
+      #    mixins = os_tpl_list #=> #<Occi::Core::Mixins>
+      #    mixins.first #=> #<Occi::Core::Mixin>
+      #
+      # @return [Occi::Core::Mixins] a collection of mixins
+      # @effects Gets status of machine images
+      def os_tpl_list
+        filters = []
+        filters << { name: 'image-type', values: ['machine'] }
+        filters << { name: 'image-id', values: @image_filtering_image_list } if IMAGE_FILTERING_POLICIES_LISTED.include?(@image_filtering_policy)
+        owners = IMAGE_FILTERING_POLICIES_OWNED.include?(@image_filtering_policy) ? [ 'self' ] : nil
+
+        ec2_images_ary = nil
+        unless ec2_images_ary = Backends::Helpers::CachingHelper.load(@dalli_cache, DALLI_OS_TPL_KEY)
+          ec2_images_ary = []
+
+          Backends::Ec2::Helpers::AwsConnectHelper.rescue_aws_service(@logger) do
+            ec2_images = if owners
+                           @ec2_client.describe_images(filters: filters, owners: owners).images
+                         else
+                           @ec2_client.describe_images(filters: filters).images
+                         end
+
+            ec2_images.each { |ec2_image| ec2_images_ary << { image_id: ec2_image[:image_id], name: ec2_image[:name] } } if ec2_images
+          end
+
+          Backends::Helpers::CachingHelper.save(@dalli_cache, DALLI_OS_TPL_KEY, ec2_images_ary)
+        end
+
+        os_tpls = Occi::Core::Mixins.new
+        ec2_images_ary.each { |ec2_image| os_tpls << os_tpl_list_mixin_from_image(ec2_image) }
+
+        os_tpls
+      end
+
+      # Gets a specific os_tpl mixin instance as Occi::Core::Mixin.
+      # Term given as an argument must match the term inside
+      # the returned Occi::Core::Mixin instance.
+      #
+      # @example
+      #    os_tpl = os_tpl_get('65d4f65adfadf-ad2f4ad-daf5ad-f5ad4fad4ffdf')
+      #        #=> #<Occi::Core::Mixin>
+      #
+      # @param term [String] OCCI term of the requested os_tpl mixin instance
+      # @return [Occi::Core::Mixin, nil] a mixin instance or `nil`
+      # @effects Gets status of a given machine image
+      def os_tpl_get(term)
+        filters = []
+        filters << { name: 'image-type', values: ['machine'] }
+        filters << { name: 'image-id', values: [term] }
+
+        Backends::Ec2::Helpers::AwsConnectHelper.rescue_aws_service(@logger) do
+          ec2_images = @ec2_client.describe_images(filters: filters).images
+          (ec2_images && ec2_images.first) ? os_tpl_list_mixin_from_image(ec2_images.first) : nil
+        end
+      end
+
+      #
+      #
+      def os_tpl_list_image_to_term(ec2_image)
+        ec2_image[:image_id]
+      end
+
+      #
+      #
+      def os_tpl_list_term_to_image_id(term)
+        term
+      end
+
+      #
+      #
+      def os_tpl_list_mixin_from_image(ec2_image)
+        depends = %w|http://schemas.ogf.org/occi/infrastructure#os_tpl|
+        term = os_tpl_list_image_to_term(ec2_image)
+        scheme = "#{@options.backend_scheme}/occi/infrastructure/os_tpl#"
+        title = ec2_image[:name] || 'unknown'
+        location = "/mixin/os_tpl/#{term}/"
+        applies = %w|http://schemas.ogf.org/occi/infrastructure#compute|
+
+        ::Occi::Core::Mixin.new(scheme, term, title, nil, depends, nil, location, applies)
+      end
+
+      # Gets platform- or backend-specific `resource_tpl` mixins which should be merged
+      # into Occi::Model of the server.
+      #
+      # @example
+      #    mixins = resource_tpl_list #=> #<Occi::Core::Mixins>
+      #    mixins.first  #=> #<Occi::Core::Mixin>
+      #
+      # @return [Occi::Core::Mixins] a collection of mixins
+      # @effects <i>none</i>: call answered from within the backend
+      def resource_tpl_list
+        @resource_tpl
+      end
+
+      # Gets a specific resource_tpl mixin instance as Occi::Core::Mixin.
+      # Term given as an argument must match the term inside
+      # the returned Occi::Core::Mixin instance.
+      #
+      # @example
+      #    resource_tpl = resource_tpl_get('65d4f65adfadf-ad2f4ad-daf5ad-f5ad4fad4ffdf')
+      #        #=> #<Occi::Core::Mixin>
+      #
+      # @param term [String] OCCI term of the requested resource_tpl mixin instance
+      # @return [Occi::Core::Mixin, nil] a mixin instance or `nil`
+      # @effects <i>none</i>: call answered from within the backend
+      def resource_tpl_get(term)
+        resource_tpl_list.to_a.select { |m| m.term == term }.first
+      end
+
+      #
+      #
+      def resource_tpl_list_itype_to_term(ec2_itype)
+        ec2_itype ? ec2_itype.gsub('.', '_') : nil
+      end
+
+      #
+      #
+      def resource_tpl_list_term_to_itype(term)
+        term ? term.gsub('_', '.') : nil
       end
 
       private
