+++ lib/backends/ec2/storage.rb
@@ -1,6 +1,6 @@
 module Backends
   module Ec2
-    module Storage
+    class Storage < Backends::Ec2::Base
       # Gets all storage instance IDs, no details, no duplicates. Returned
       # identifiers must correspond to those found in the occi.core.id
       # attribute of Occi::Infrastructure::Storage instances.
@@ -239,6 +239,14 @@ module Backends
         end
 
         true
+      end
+
+      # Returns a collection of custom mixins introduced (and specific for)
+      # the enabled backend. Only mixins and actions are allowed.
+      #
+      # @return [Occi::Collection] collection of extensions (custom mixins and/or actions)
+      def storage_get_extensions
+        read_extensions 'storage', @options.model_extensions_dir
       end
 
       private
